requirements.txt
pandas
pytest
pdfplumber
google-generativeai
langchain
langgraph
tiktoken
requests
